pyproject.toml
[project]
name = "ytvideosummarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.127.0",
    "google-genai>=1.56.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-qdrant>=0.1.0",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.10.0",
    "uvicorn>=0.40.0",
    "youtube-transcript-api>=1.2.3",
]
